requirements.txt
llama-stack
faiss-cpu
langchain
PyPDF2
transformers
torch
sentence-transformers
